requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
cryptography>=41.0.0
plaid-python>=21.0.0
